Updated Version
Version was changed to have smooth collision to compensate for integration due to one team member creating their own substantially changed template. Idea was to integrate their template as best as possible. Collectively decided to include the template in the presentation, but not in the final game. Team member could not effectively communicate their methodology to the group. Template had substantial constraints.